requirements.txt
python-dotenv
line-bot-sdk
google-generativeai==0.3.2
flask
